Q4 form scoring project/backend/response_scorer/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
PyJWT==2.*
Flask==3.*
